remove KickstarterUITests scheme and target, moved tests from KickstarterUITests to UIViewController+URLTests. Next step is update ci.
